requirements.txt
Babel==1.3
Flask==0.10.1
Flask-Babel==0.9
Flask-WTF==0.9.5
Jinja2==2.7.2
Markdown==2.4.1
MarkupSafe==0.23
TwitterAPI==2.1.13
WTForms==1.0.5
Werkzeug==0.9.4
flup==1.0.2
itsdangerous==0.24
oauthlib==0.6.1
pytz==2014.3
requests==2.3.0
requests-oauthlib==0.4.0
speaklater==1.3
wsgiref==0.1.2
